# .kilo/worktrees/equinox-ocean/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JurisLab API",
  "description": "Multi-agent legal simulation and document review platform (educational).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/documents/upload": {
   "post": {
    "tags": [
     "documents"
    ],
    "summary": "Upload Document",
    "operationId": "upload_document_api_v1_documents_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Upload Document Api V1 Documents Upload Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/documents": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "List Documents",
    "operationId": "list_documents_api_v1_documents_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Documents Api V1 Documents Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{document_id}": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Get Document",
    "operationId": "get_document_api_v1_documents__document_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Document Api V1 Documents  Document Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "documents"
    ],
    "summary": "Delete Document",
    "operationId": "delete_document_api_v1_documents__document_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/documents/{document_id}/detail": {
   "get": {
    "tags": [
     "documents"
    ],
    "summary": "Document Detail",
    "operationId": "document_detail_api_v1_documents__document_id__detail_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Document Detail Api V1 Documents  Document Id  Detail Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/reviews/{review_id}": {
   "get": {
    "tags": [
     "reviews"
    ],
    "summary": "Get Review",
    "operationId": "get_review_api_v1_reviews__review_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "review_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Review Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Review Api V1 Reviews  Review Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/exports/reviews/{review_id}": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Export Review",
    "operationId": "export_review_api_v1_exports_reviews__review_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "review_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Review Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Export Review Api V1 Exports Reviews  Review Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/exports/case_studies/{case_study_id}": {
   "get": {
    "tags": [
     "exports"
    ],
    "summary": "Export Case Study Route",
    "operationId": "export_case_study_route_api_v1_exports_case_studies__case_study_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "case_study_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Study Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Export Case Study Route Api V1 Exports Case Studies  Case Study Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Body_upload_document_api_v1_documents_upload_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     },
     "title": {
      "type": "string",
      "title": "Title",
      "default": ""
     },
     "doc_type": {
      "type": "string",
      "title": "Doc Type",
      "default": "contract"
     },
     "jurisdiction": {
      "type": "string",
      "title": "Jurisdiction",
      "default": ""
     },
     "domain": {
      "type": "string",
      "title": "Domain",
      "default": ""
     },
     "effective_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Effective Date"
     },
     "version": {
      "type": "string",
      "title": "Version",
      "default": "1.0"
     },
     "source_status": {
      "type": "string",
      "title": "Source Status",
      "default": "verified"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_document_api_v1_documents_upload_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}