# backend/app/api/v1/exports.py
"""Compliance-enabled export endpoints: review reports and case-study packs (PRD §8.6, §15)."""

from __future__ import annotations

from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.orm import Session

from app.core.audit import audit
from app.core.deps import CurrentUser, enforce_export, require_module, require_tenant
from app.core.ratelimit import rate_limit
from app.db.session import get_db
from app.models import CaseStudy, DocumentReview

router = APIRouter(prefix="/exports", tags=["exports"])


@router.get("/reviews/{review_id}")
def export_review(
    review_id: str,
    user: CurrentUser = Depends(require_module("review", "report.export")),
    db: Session = Depends(get_db),
    _: None = Depends(rate_limit("exports.get")),
) -> dict:
    review = (
        db.query(DocumentReview).filter(DocumentReview.id == review_id).one_or_none()
    )
    if review is None or review.organization_id != user.organization_id:
        raise HTTPException(status_code=404, detail="review not found")
    enforce_export(
        user,
        "review_report",
        {
            "contains_plain_language": bool(
                review.findings and any(f.plain_language for f in review.findings)
            ),
            "status": review.status,
        },
    )
    audit(
        db,
        action="exported",
        actor=user.id,
        organization_id=user.organization_id,
        resource_type="review",
        resource_id=review_id,
        detail={"kind": "review_report"},
    )
    return {
        "kind": "review_report",
        "review_id": review.id,
        "document_id": review.document_id,
        "markdown": review.markdown_report,
        "filename": f"review_{review.id[:8]}.md",
        "disclaimer": "Generated by JurisFlow for educational use only; not legal advice.",
    }


@router.get("/case_studies/{case_study_id}")
def export_case_study_route(
    case_study_id: str,
    user: CurrentUser = Depends(require_module("sim", "case_study.export")),
    db: Session = Depends(get_db),
    _: None = Depends(rate_limit("exports.get")),
) -> dict:
    study = db.query(CaseStudy).filter(CaseStudy.id == case_study_id).one_or_none()
    if study is None or study.organization_id != user.organization_id:
        raise HTTPException(status_code=404, detail="case study not found")
    require_tenant(study.organization_id, user)
    enforce_export(
        user, "case_study", {"contains_grounded_citations": bool(study.citations or [])}
    )
    audit(
        db,
        action="exported",
        actor=user.id,
        organization_id=user.organization_id,
        resource_type="case_study",
        resource_id=case_study_id,
        detail={"kind": "case_study"},
    )
    return {
        "kind": "case_study",
        "case_study_id": study.id,
        "simulation_id": study.simulation_id,
        "markdown": study.markdown_content,
        "filename": f"case_study_{study.id[:8]}.md",
        "citations": study.citations or [],
        "disclaimer": "Generated by JurisFlow for educational use only; not legal advice.",
    }